+++ scripts/build-aws-gateway-function.sh
@@ -1,0 +1,1 @@
+cargo lambda build -p email_sender_aws_gateway --release --arm64 --output-format zip