+++ scripts/upgrade_canister.sh
@@ -4,7 +4,7 @@ IDENTITY=$1
 IC_URL=$2
 CANISTER_ID=$3
 AWS_REGION=$4
-AWS_TARGET_ARN=$5
+AWS_FUNCTION_URL=$5
 AWS_ACCESS_KEY=$6
 AWS_SECRET_KEY=$7
 
@@ -18,6 +18,6 @@ cargo run \
     --ic-url $IC_URL \
     --canister-id $CANISTER_ID \
     --aws-region $AWS_REGION \
-    --aws-target-arn $AWS_TARGET_ARN \
+    --aws-function-url $AWS_FUNCTION_URL \
     --aws-access-key $AWS_ACCESS_KEY \
     --aws-secret-key $AWS_SECRET_KEY \